Crate asset_hub_westend_runtime

Asset Hub Westend Runtime

Testnet for Asset Hub Polkadot.

Modules

api

governance

Governance configurations for the Asset Hub runtime.

xcm_config

Macros

add_benchmarks

Calls add_benchmark with all configs from define_benchmarks and passes the first two parameters on.

list_benchmarks

Calls list_benchmark with all configs from define_benchmarks and passes the first two parameters on.

Structs

AnnouncementDepositBase

AnnouncementDepositFactor

ApprovalDeposit

AssetAccountDeposit

AssetConversionOrigin

AssetConversionPalletId

AssetConversionTxHelper

AssetDeposit

AssetRewardsPalletId

AssetsStringLimit

BaseDeliveryFee

The base fee for the message delivery fees.

CodeHashLockupDepositPercent

DeleteUndecodableStorage

Asset Hub Westend has some undecodable storage, delete it. See https://github.com/paritytech/polkadot-sdk/issues/2241 for more info.

DepositBase

DepositFactor

DepositPerByte

DepositPerItem

EthExtraImpl

Default extensions applied to Ethereum transactions.

ExistentialDeposit

FeeAssetId

The asset ID for the asset that we use to pay for message delivery fees.

ForeignAssetsApprovalDeposit

ForeignAssetsAssetAccountDeposit

ForeignAssetsAssetDeposit

ForeignAssetsAssetsStringLimit

ForeignAssetsMetadataDepositBase

ForeignAssetsMetadataDepositPerByte

IndexDeposit

Deposit for an index in the indices pallet.

InitStorageVersions

Migration to initialize storage versions for pallets added after genesis.

LiquidityWithdrawalFee

MaxPending

MaxProxies

MaxSignatories

MaximumSchedulerWeight

MbmServiceWeight

MessageQueueServiceWeight

MetadataDepositBase

Key = 32 bytes, Value = 36 bytes (32+1+1+1+1)

MetadataDepositPerByte

MigController

MigrationMaxKeyLen

MigrationSignedDepositBase

MigrationSignedDepositPerItem

MinVestedTransfer

NewAssetName

NewAssetSymbol

NftFractionalizationPalletId

NftsAttributeDepositBase

NftsCollectionDeposit

NftsDepositPerByte

NftsItemDeposit

NftsMaxDeadlineDuration

NftsMetadataDepositBase

NftsPalletFeatures

Offset

PalletAssetRewardsBenchmarkHelper

PalletInfo

Provides an implementation of PalletInfo to provide information about the pallet setup in the runtime.

Period

PotId

PreimageBaseDeposit

PreimageByteDeposit

PreimageHoldReason

ProxyDepositBase

ProxyDepositFactor

RelayOrigin

ReservedDmpWeight

ReservedXcmpWeight

RewardsPoolCreationHoldReason

RootMigController

Runtime

RuntimeApi

RuntimeApiImpl

Implements all runtime apis for the client side.

RuntimeBlockLength

RuntimeBlockWeights

RuntimeGenesisConfig

RuntimeOrigin

The runtime origin type representing the origin of a call.

SS58Prefix

SessionKeys

SessionLength

StakePoolCreationDeposit

StakingPot

TransactionByteFee

Relay Chain TransactionByteFee / 10

UniquesAttributeDepositBase

UniquesCollectionDeposit

UniquesDepositPerByte

UniquesItemDeposit

UniquesMetadataDepositBase

UnvestedFundsAllowedWithdrawReasons

Version

Enums

OriginCaller

ProxyType

The type used to represent the kinds of proxying allowed.

RuntimeCall

The aggregated runtime call type.

RuntimeError

RuntimeEvent

RuntimeFreezeReason

A reason for placing a freeze on funds.

RuntimeHoldReason

A reason for placing a hold on funds.

RuntimeLockId

An identifier for each lock placed on funds.

RuntimeSlashReason

A reason for slashing funds.

RuntimeTask

An aggregation of all Task enums across all pallets included in the current runtime.

RuntimeViewFunction

Runtime query type.

Constants

RUNTIME_API_VERSIONS

VERSION

WASM_BINARY

WASM_BINARY_BLOATY

WASM_BINARY_PATH

Traits

BuildStorage

Complex storage builder stuff.

Functions

native_version

The version information used to identify this runtime when compiled natively.

Type Aliases

Address

The address format for describing accounts.

AhOps

AllPalletsWithSystem

All pallets included in the runtime as a nested tuple of types.

AllPalletsWithoutSystem

All pallets included in the runtime as a nested tuple of types. Excludes the System pallet.

AssetConversion

AssetConversionMigration

AssetRate

AssetRewards

AssetTxPayment

Assets

AssetsConfig

AssetsForceOrigin

AssetsFreezer

AssetsFreezerInstance

Aura

AuraConfig

AuraExt

AuraExtConfig

Authorship

Balances

BalancesConfig

Block

Block type as expected by this runtime.

BlockId

BlockId type as expected by this runtime.

CollatorSelection

CollatorSelectionConfig

CollatorSelectionUpdateOrigin

ConvictionVoting

CumulusXcm

DelegatedStaking

Executive

Executive: handles dispatch to the various modules.

FastUnstake

ForeignAssets

ForeignAssetsConfig

ForeignAssetsFreezer

ForeignAssetsFreezerInstance

ForeignAssetsInstance

Assets managed by some foreign location. Note: we do not declare a ForeignAssetsCall type, as this type is used in proxy definitions. We assume that a foreign location would not want to set an individual, local account as a proxy for the issuance of their assets. This issuance should be managed by the foreign location’s governance.

Indices

IndicesConfig

LocalAndForeignAssets

Union fungibles implementation for Assets and ForeignAssets.

LocalAndForeignAssetsFreezer

Union fungibles implementation for AssetsFreezer and ForeignAssetsFreezer.

MessageQueue

Migrations

Migrations to apply on runtime upgrade.

MultiBlockElection

MultiBlockElectionSigned

MultiBlockElectionUnsigned

MultiBlockElectionVerifier

MultiBlockElectionVerifierConfig

MultiBlockMigrations

Multisig

NativeAndAllAssets

Union fungibles implementation for PoolAssets and NativeAndNonPoolAssets.

NativeAndAllAssetsFreezer

Union fungibles implementation for PoolAssetsFreezer and NativeAndNonPoolAssetsFreezer.

NativeAndNonPoolAssets

Union fungibles implementation for LocalAndForeignAssets and Balances.

NativeAndNonPoolAssetsFreezer

Union fungibles implementation for LocalAndForeignAssetsFreezer and Balances.

NftFractionalization

Nfts

NominationPools

NominationPoolsConfig

Origins

ParachainInfo

ParachainInfoConfig

ParachainSystem

ParachainSystemConfig

PolkadotXcm

PolkadotXcmConfig

PoolAssets

PoolAssetsConfig

PoolAssetsFreezer

PoolAssetsFreezerInstance

PoolAssetsInstance

PoolIdToAccountId

Preimage

PriceForSiblingParachainDelivery

Proxy

Referenda

Revive

ReviveConfig

Scheduler

Session

SessionConfig

SignedBlock

A Block signed with a Justification

SnowbridgeSystemFrontend

Staking

StakingConfig

StakingRcClient

StateTrieMigration

Sudo

SudoConfig

System

SystemConfig

Timestamp

ToRococoXcmRouter

ToRococoXcmRouterInstance

XCM router instance to BridgeHub with bridging capabilities for Rococo global consensus with dynamic fees and back-pressure.

TransactionPayment

TransactionPaymentConfig

Treasury

TreasuryConfig

TrustBackedAssetsInstance

TxExtension

The extension to the basic transaction logic.

UncheckedExtrinsic

Unchecked extrinsic type as expected by this runtime.

Uniques

Utility

Vesting

VestingConfig

VoterList

WeightReclaim

Whitelist

XcmpQueue