asset_hub_westend_runtime

Module staking

Structs

AutoRebagNumber

BagThresholds

BailoutGraceRatio

BenchElectionBounds

BondingDuration

DapPalletId

DelegatedStakingPalletId

DepositBase

DepositPerPage

EjectGraceRatio

IssuanceCadence

Minimum time (ms) between issuance drips. 60s = drip at most once per minute.

IssuanceCurve

Westend inflation curve for DAP.

KeysMessageToXcm

MaxBackersPerWinner

In each page of the election, we allow up to all of the nominators of that page to be present.

MaxBackersPerWinnerFinal

Total number of backers per winner across all pages.

MaxControllersInDeprecationBatch

MaxElapsedPerDrip

Safety ceiling (ms) for elapsed time in a single drip. Prevents over-minting after stalls.

MaxElectingVoters

Compatible with Polkadot, we allow up to 22_500 nominators to be considered for election

MaxEraDuration

MaxExposurePageSize

Size of the exposures. This should be small enough to make the reward payouts feasible.

MaxNominations

MaxPointsToBalance

MaxPruningItems

MaxSubmissions

MaxValidatorSet

Maximum number of validators that we may want to elect. 1000 is the end target.

MaxWinnersPerPage

In each page, we may observe up to all of the validators.

MinerMaxLength

Miner transaction can fill up to 75% of the block size.

MinerPages

MinerTxPriority

Priority of the offchain miner transactions.

NominatorFastUnbondDuration

NposCompactSolution16

OffchainRepeat

Try and run the OCW miner 4 times during the unsigned phase.

OnChainConfig

Pages

Number of election pages that we operate upon. 32 * 6s block = 192s = 3.2min snapshots

PoolsPalletId

RelayChainSessionKeys

Generated by impl_opaque_keys!.

RelayLocation

RelaySessionDuration

Duration of a relay session in our blocks. Needs to be hardcoded per-runtime.

RemoteKeysExecutionWeight

Conservative RC execution cost for set/purge keys operations. Intentionally ~2-3x of benchmarked values to avoid undercharging if RC weights increase. Slight overpaymennt is the price we pay for maintainability here.

RewardBase

SessionsPerEra

SignedPhase

SignedValidationPhase

validate up to 4 signed solution. Each solution.

SlashDeferDuration

SlashRewardFraction

StakingPotsPalletId

StakingStakerRewardKind

StakingXcmToRelayChain

TargetSnapshotPerBlock

Number of validators per page of the snapshot.

UnsignedPhase

ValidatorSetToXcm

VoterSnapshotPerBlock

Number of nominators per page of the snapshot, and consequently number of backers in the solution. Uses ceiling division so that VoterSnapshotPerBlock * Pages >= MaxElectingVoters holds for any configured values.

WestendStakingMiner

Enums

AhClientCalls

RelayChainRuntimePallets

Type Aliases

SignedPayload

The payload being signed in transactions.

UncheckedExtrinsic

Unchecked extrinsic type as expected by this runtime.